--- src/PenLayer.js
@@ -1,131 +1,0 @@
-const twgl = require('twgl.js');
-
-const vertexShaderText = require('./shaders/sprite.vert');
-const fragmentShaderText = require('./shaders/sprite.frag');
-
-
-class PenLayer {
-    /**
-     * Create a Pen Layer.
-     * @param {WebGLRenderingContext} gl - The WebGL rendering context to use.
-     */
-    constructor (gl) {
-        /** @type {WebGLRenderingContext} */
-        this._gl = gl;
-
-        /** @type {HTMLCanvasElement} */
-        this._canvas = document.createElement('canvas');
-
-        /** @type {boolean} */
-        this._textureDirty = false;
-
-        /** @type {DrawObject[]} */
-        this._drawObjects = [{
-            type: gl.TRIANGLE_STRIP,
-            programInfo: twgl.createProgramInfo(gl, [vertexShaderText, fragmentShaderText]),
-            bufferInfo: twgl.createBufferInfoFromArrays(gl, {
-                a_position: {
-                    numComponents: 2,
-                    data: [
-                        -0.5, -0.5,
-                        0.5, -0.5,
-                        -0.5, 0.5,
-                        0.5, 0.5
-                    ]
-                },
-                a_texCoord: {
-                    numComponents: 2,
-                    data: [
-                        1, 0,
-                        0, 0,
-                        1, 1,
-                        0, 1
-                    ]
-                }
-            }),
-            uniforms: {
-                /** @type {WebGLTexture} */
-                u_skin: null,
-
-                /** @type {module:twgl/m4.Mat4} */
-                u_modelMatrix: null,
-
-                /** @type {module:twgl/m4.Mat4} */
-                u_projectionMatrix: null
-            }
-        }];
-    }
-
-    /**
-     * Resize the pen layer's working area.
-     * @param {int} width - The new width of the working area.
-     * @param {int} height - The new height of the working area.
-     */
-    resize (width, height) {
-        const gl = this._gl;
-        this._canvas.width = width;
-        this._canvas.height = height;
-        this._drawObjects[0].uniforms.u_modelMatrix = twgl.m4.scaling([-width, -height, 1]);
-        this._texture = twgl.createTexture(
-            gl,
-            {
-                auto: true,
-                mag: gl.NEAREST,
-                min: gl.NEAREST,
-                wrap: gl.CLAMP_TO_EDGE,
-                src: this._canvas
-            }
-        );
-        this._textureDirty = true;
-
-        const ctx = this._canvas.getContext('2d');
-        ctx.strokeStyle = 'red';
-        ctx.beginPath();
-        ctx.moveTo(-2, 0);
-        ctx.lineTo(width - 2, height);
-        ctx.stroke();
-        ctx.strokeStyle = 'green';
-        ctx.beginPath();
-        ctx.moveTo(0, 0);
-        ctx.lineTo(width + 0, height);
-        ctx.stroke();
-        ctx.strokeStyle = 'blue';
-        ctx.beginPath();
-        ctx.moveTo(2, 0);
-        ctx.lineTo(width + 2, height);
-        ctx.stroke();
-        ctx.strokeStyle = 'purple';
-        ctx.moveTo(2, height / 10);
-        ctx.lineTo(width / 10, height / 5);
-        ctx.stroke();
-    }
-
-    /**
-     * Draw the pen layer onto the current frame buffer.
-     * @param {module:twgl/m4.Mat4} projectionMatrix The projection matrix to use.
-     */
-    draw (projectionMatrix) {
-        if (this._texture) {
-            const gl = this._gl;
-            if (this._textureDirty) {
-                this._textureDirty = false;
-                gl.bindTexture(gl.TEXTURE_2D, this._texture);
-                gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, this._canvas);
-            }
-            this._drawObjects[0].uniforms.u_projectionMatrix = projectionMatrix;
-            twgl.drawObjectList(gl, this._drawObjects);
-        }
-    }
-
-    /** @returns {WebGLTexture} The current pen layer texture. */
-    get _texture () {
-        return this._drawObjects[0].uniforms.u_skin;
-    }
-
-    /** @param {WebGLTexture} newTexture - The new pen layer texture. */
-    set _texture (newTexture) {
-        this._drawObjects[0].uniforms.u_skin = newTexture;
-    }
-}
-
-module.exports = PenLayer;
